requirements.txt
django>=1.6,<1.7
django-autocomplete-light
django-photologue
South==0.7.6
feedparser
lxml
django-extensions
django-guardian
django-timedeltafield
python-ldap
